requirements.txt
beautifulsoup4
pdfminer.six
